Move question about job market to the end of the Rust at work section
Before, we had a long section about Rust at work, then we asked about the user's opinion of Rust and their worries for its future, and then we went back to the job market. I think that it makes sense to ask about the job market right after asking about Rust at work. Note that this question is applicable even if the person is not currently employed.